Changed package name to kmdlib to avoid confusion. More stubs added to functions in kademlia.go
