+++ main/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"Kademlia---P2P-DFS/kademlia"
+	"Kademlia---P2P-DFS/kdmlib"
 	"fmt"
 )
 
 func main() {
-	contact := kademlia.NewContact(kademlia.NewKademliaID("FFFFFFFF00000000000000000000000000000000"), "localhost:8000")
+	contact := kdmlib.NewContact(kdmlib.NewKademliaID("FFFFFFFF00000000000000000000000000000000"), "localhost:8000")
 	fmt.Print(contact.String())
 }
